Move validators to question files and simplify error handling
